refine

+++ app/src/main/java/org/wikipedia/page/PageFragment.kt
@@ -187,7 +187,7 @@ class PageFragment : Fragment(), BackPressedHandler, CommunicationBridge.Communi
     override val isPreview get() = false
     override val referencesGroup get() = references?.referencesGroup
     override val selectedReferenceIndex get() = references?.selectedIndex ?: 0
-    override var messageCardHeight: Int = 0
+    override var messageCardHeight: Float = 0f
 
     lateinit var sidePanelHandler: SidePanelHandler
     lateinit var shareHandler: ShareHandler
